+++ src/command_modules/azure-cli-acr/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+2.1.4
++++++
+* Provide a workaround for runtime operations without ARM requests.
+* Exclude version control files (eg, .git, .gitignore) from uploaded tar by default in build command.
+
 2.1.3
 +++++
 * Add content-trust policy commands.