+++ src/command_modules/azure-cli-batch/HISTORY.rst
@@ -5,7 +5,7 @@ Release History
 
 3.3.3
 +++++
-* Minor fixes
+* Update Batch Management SDK dependency
 
 3.3.2
 +++++