+++ src/command_modules/azure-cli-monitor/HISTORY.rst
@@ -5,7 +5,8 @@ Release History
 
 0.2.3
 +++++
-* Minor fixes
+* Added `monitor metrics alert` commands for near-realtime metric alerts.
+* Deprecated `monitor alert` commands.
 
 0.2.2
 +++++