+++ src/command_modules/azure-cli-role/HISTORY.rst
@@ -2,6 +2,11 @@
 
 Release History
 ===============
+2.1.4
+++++++
+* Minor fixes.
+
+
 2.1.3
 ++++++
 * role assignment: fix a recent regression that principalName is missing